Fixes bug in xs3_clip_vect_s32(). Adds C implementations for xs3_clip_vect_s16() and xs3_clip_vect_s32(). Fixes bug in HR_C32() macro. Gets several unit tests working.

+++ lib_xs3_math/api/low/xs3_math_util_low.h
@@ -54,7 +54,7 @@ extern "C" {
  * 
  * \sa cls()
  */
-#define CLS_C32(X)      (XS3_MIN(CLS_S32(((int16_t)(X).re)), CLS_S32(((int16_t)(X).im))))
+#define CLS_C32(X)      (XS3_MIN(CLS_S32(((int32_t)(X).re)), CLS_S32(((int32_t)(X).im))))
